@@ -4,18 +4,17 @@
   "private": true,
   "author": "shi9uma <https://github.com/shi9uma>",
   "description": "A Minecraft Java Edition Launcher with Vue",
+  "type": "module",
   "main": "index.js",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "electron:serve": "vue-cli-service build && electron .",
-    "electron:build": "vue-cli-service build && electron-builder build --windows"
+    "dev": "vite",
+    "electron:dev": "vite build && electron .",
+    "build": "vite build && electron-builder"
   },
   "build": {
     "appId": "github.shi9uma.kisara-launcher",
     "productName": "Kisara-Launcher",
-    "copyright": "shiguma",
+    "copyright": "shi9uma",
     "directories": {
       "output": "dist_electron"
     },
@@ -30,39 +29,12 @@
     }
   },
   "dependencies": {
-    "core-js": "^3.8.3",
-    "vue": "^3.2.13"
+    "vue": "^3.4.21"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
+    "@vitejs/plugin-vue": "^5.0.4",
     "electron": "^29.1.5",
     "electron-builder": "^24.13.3",
-    "electron-builder-squirrel-windows": "^24.13.3",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/vue3-essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "@babel/eslint-parser"
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead",
-    "not ie 11"
-  ]
+    "vite": "^5.2.0"
+  }
 }
