@@ -1,29 +1,34 @@
-<template>
-  <!-- <img alt="Vue logo" src="./assets/logo.png"> -->
-  <SumCalculator />
-  <!-- <HelloWorld msg="Welcome to Your Vue.js App"/> -->
-</template>
-
-<script>
-// import HelloWorld from './components/HelloWorld.vue'
-import SumCalculator from './components/SumCalculator.vue';
-
-export default {
-  name: 'App',
-  components: {
-    // HelloWorld,
-    SumCalculator
-  }
-}
+<script setup>
+import HelloWorld from './components/HelloWorld.vue'
+import sum_test from './components/sum_test.vue';
 </script>
 
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
+<template>
+  <div>
+    <a href="https://vitejs.dev" target="_blank">
+      <img src="/vite.svg" class="logo" alt="Vite logo" />
+    </a>
+    <a href="https://vuejs.org/" target="_blank">
+      <img src="./assets/vue.svg" class="logo vue" alt="Vue logo" />
+    </a>
+  </div>
+  <HelloWorld msg="Vite + Vue" />
+  <sum_test />
+</template>
+
+<style scoped>
+.logo {
+  height: 6em;
+  padding: 1.5em;
+  will-change: filter;
+  transition: filter 300ms;
+}
+
+.logo:hover {
+  filter: drop-shadow(0 0 2em #646cffaa);
+}
+
+.logo.vue:hover {
+  filter: drop-shadow(0 0 2em #42b883aa);
 }
 </style>
